refactor: [Common-Web] GatewayAuthFilter와 SecurityErrorHandlers를 common-web으로 추출 (#164)
- common-web에 GatewayAuthFilter 추가 (@Component 제거, 이중 등록 방지)
- common-web에 SecurityErrorHandlers object 추가 (401/403 팩토리 함수)
- event-service, queue-service SecurityConfig에서 인라인 인스턴스화로 전환
- queue-service 의존성: common-core → common-web으로 변경
- 로컬 중복 GatewayAuthFilter 2개 삭제
